Conflicts:
	x86InfoAndControl_MSVC/x86InfoAndControl/x86InfoAndControl.vcproj.Stefan-PC.Stefan.user

+++ Controller/CPU-related/Intel/Core/Core.cpp
@@ -4,5 +4,6 @@
  *  Created on: 28.09.2010
  *      Author: sgebauer
  */
+#include <stdint.h> //uint32_t
 
-  unsigned long g_dwValue1, g_dwValue2 ;
+uint32_t g_dwValue1, g_dwValue2 ;
